+++ weblate/checks/models.py
@@ -79,6 +79,7 @@ class WeblateChecksConf(AppConf):
         'weblate.checks.markup.MarkdownSyntaxCheck',
         'weblate.checks.markup.URLCheck',
         'weblate.checks.placeholders.PlaceholderCheck',
+        'weblate.checks.placeholders.RegexCheck',
         'weblate.checks.source.OptionalPluralCheck',
         'weblate.checks.source.EllipsisCheck',
         'weblate.checks.source.MultipleFailingCheck',