+++ weblate/checks/placeholders.py
@@ -18,9 +18,13 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
+from __future__ import unicode_literals
+
 import re
 
 from django.utils.translation import ugettext_lazy as _
+from django.utils.safestring import mark_safe
+from django.utils.html import escape
 
 from weblate.checks.base import TargetCheckParametrized
 
@@ -28,6 +32,11 @@ from weblate.checks.base import TargetCheckParametrized
 @staticmethod
 def parse_placeholders(val):
     return val.split(":")
+
+
+@staticmethod
+def parse_regex(val):
+    return re.compile(val)
 
 
 class PlaceholderCheck(TargetCheckParametrized):
@@ -51,3 +60,33 @@ class PlaceholderCheck(TargetCheckParametrized):
         for match in re.finditer(regexp, source):
             ret.append((match.start(), match.end(), match.group()))
         return ret
+
+
+class RegexCheck(TargetCheckParametrized):
+    check_id = "regex"
+    default_disabled = True
+    name = _("Regular expression")
+    description = _("Translation does not match regular expression:")
+    severity = "danger"
+    param_type = parse_regex
+
+    def check_target_params(self, sources, targets, unit, value):
+        return any(not value.findall(target) for target in targets)
+
+    def check_highlight(self, source, unit):
+        if self.should_skip(unit):
+            return []
+        ret = []
+
+        regex = self.get_value(unit)
+
+        for match in regex.finditer(source):
+            ret.append((match.start(), match.end(), match.group()))
+        return ret
+
+    def get_description(self, unit):
+        regex = self.get_value(unit)
+        return mark_safe('{} <code>{}</code>'.format(
+            escape(self.description),
+            escape(regex.pattern)
+        ))